// src/services/zapper/types/models/network.ts
export enum Network {
  ETHEREUM_MAINNET = 'ethereum',
  POLYGON_MAINNET = 'polygon',
  OPTIMISM_MAINNET = 'optimism',
  // GNOSIS_MAINNET = 'gnosis',
  BINANCE_SMART_CHAIN_MAINNET = 'binance-smart-chain',
  FANTOM_OPERA_MAINNET = 'fantom',
  AVALANCHE_MAINNET = 'avalanche',
  ARBITRUM_MAINNET = 'arbitrum',
  // CELO_MAINNET = 'celo',
  // HARMONY_MAINNET = 'harmony',
  // MOONRIVER_MAINNET = 'moonriver',
  // BITCOIN_MAINNET = 'bitcoin',
  CRONOS_MAINNET = 'cronos',
  AURORA_MAINNET = 'aurora',
  // EVMOS_MAINNET = 'evmos',
}

export type ALL_NETWORKS = 'all-networks'